requirements.txt
opencv-python == 4.5.4.58
numpy == 1.21.3
torchvision == 0.11.3
torch == 1.10.2
pillow == 7.0.0
